requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
qrcode==8.2
Pillow>=12,<13
gunicorn==23.0.0
